fix: Add build constraint to futex_debug_test.go for 386 arch
On 386 architecture, syscall.Timespec.Sec and Nsec are int32, not int64.
This test uses raw syscall interfaces with 64-bit timestamps, so restrict
it to linux amd64/arm64 architectures where the types match.

+++ internal/transport/futex_debug_test.go
@@ -1,3 +1,5 @@
+//go:build linux && (amd64 || arm64)
+
 /*
  * Copyright 2024 gRPC authors.
  *
@@ -37,8 +39,8 @@ func TestFutexDirect(t *testing.T) {
 	// Manual futex call with timeout
 	var ts syscall.Timespec
 	timeoutNs := int64(100 * 1000 * 1000) // 100ms
-	ts.Sec = timeoutNs / 1e9
-	ts.Nsec = timeoutNs % 1e9
+	ts.Sec = int64(timeoutNs / 1e9)
+	ts.Nsec = int64(timeoutNs % 1e9)
 
 	t.Logf("Direct futex test: addr=%p, val=42, timeout=%d.%09d", addr, ts.Sec, ts.Nsec)
 
@@ -53,8 +55,8 @@ func TestFutexDirect(t *testing.T) {
 		if timeoutNs < 0 {
 			timeoutNs = 0
 		}
-		ts.Sec = timeoutNs / 1e9
-		ts.Nsec = timeoutNs % 1e9
+		ts.Sec = int64(timeoutNs / 1e9)
+		ts.Nsec = int64(timeoutNs % 1e9)
 
 		r1, r2, errno := syscall.RawSyscall6(
 			syscall.SYS_FUTEX,
